chore and refactor: ajustes nos nome no sidebarheader e importação de interfaces
