+++ src/components/Sidebar/NavItem.tsx
@@ -1,18 +1,10 @@
-import { ReactNode } from 'react';
 import { Flex, Link as ChakraLink, Icon } from '@chakra-ui/react';
 import { useNavigate, useLocation } from 'react-router-dom'; // 1. Import useLocation
-
-interface NavItemProps {
-  icon: React.ElementType;
-  children: ReactNode;
-  path?: string;
-  mt?: number;
-  onClick?: () => void;
-}
+import { NavItemProps } from '../../interface/SideBarInterface';
 
 const NavItem = ({ icon, children, path, mt, onClick }: NavItemProps) => {
   const navigate = useNavigate();
-  const location = useLocation(); // 2. Get the current location
+  const location = useLocation(); 
   
   const handleClick = () => {
     if (onClick) {
@@ -23,7 +15,7 @@ const NavItem = ({ icon, children, path, mt, onClick }: NavItemProps) => {
     }
   };
 
-  const isActive = location.pathname === path; // 3. Check if the path is active
+  const isActive = location.pathname === path; 
 
   return (
     <ChakraLink style={{ textDecoration: 'none' }} _focus={{ boxShadow: 'none' }} onClick={handleClick}>
@@ -35,8 +27,8 @@ const NavItem = ({ icon, children, path, mt, onClick }: NavItemProps) => {
         borderRadius="lg"
         role="group"
         cursor="pointer"
-        bg={isActive ? '#805AD5' : 'transparent'} // 4. Apply background if active
-        color={isActive ? 'white' : 'inherit'} // 5. Apply text color if active
+        bg={isActive ? '#805AD5' : 'transparent'} 
+        color={isActive ? 'white' : 'inherit'} 
         _hover={{
           bg: '#805AD5',
           color: 'white',
